requirements.txt
Flask
flask-restx
mysql-connector-python
Flask-Bcrypt
PyJWT
